refactor(ui): extract Channels/Scenes dialog bodies into reusable panels
Phase 1 of the notebook-tabs main-UI redesign (direction B). Pure
refactor — no behaviour change:

- ui/panels/channels_panel.py — ChannelsPanel(wx.Panel) holds the full
  channel-manager UI/logic (map switch, list, edit form, store-VFO,
  import/export). bw_label() moves here.
- ui/panels/scenes_panel.py — ScenesPanel(wx.Panel) holds the full
  band-plan UI/logic.
- ChannelsDialog / ScenesDialog become thin frames hosting the panel +
  a Close button + ESC handling. Same constructor signatures, so
  main_window is untouched.

The panels can now be embedded as notebook tabs in phase 2. The VFO
panel extraction is deferred to phase 2, where main_window's layout is
restructured (its freq/step/mode/bw widgets are entangled with the
window and the cursor controller).

Dialogs verified to construct via the panels (40-channel CB map,
13 scenes). Full suite: 93 passed.
