+++ ui/panels/__init__.py
@@ -1,0 +1,6 @@
+"""ui.panels — reusable main-window/dialog content panels.
+
+These wx.Panel subclasses hold the working UI for channels, scenes, and
+VFO tuning so the same controls can be hosted either in a modal dialog or
+as a notebook tab in the main window.
+"""
